# swagger.yaml
swagger: "2.0"
info:
  description: "Uma solução simples e funcional para quem precisa de ajuda para começar a gerir seu negócio de maneira eficiente!"
  version: "1.0.0"
  title: "HopeAPI Docs"
  contact:
    email: "nilsen.loli.emily@gmail.com"
host: "hope-api-trybe.herokuapp.com"
basePath: "/"
...
schemes:
- "https"
paths:
...
  /sales/{saleId}:
    get:
      tags:
      - "Vendas"
      summary: "Encontrar venda pelo ID"
      description: "Retorna apenas uma venda"
      operationId: "getSaleById"
      produces:
      - "application/json"
      parameters:
      - name: "saleId"
        in: "path"
        description: "ID da venda desejada"
        required: true
        type: "integer"
        format: "int64"
      responses:

        "200":
          description: "Retorna a venda selecionada"
          schema:
            $ref: "#/definitions/ModeloDeRespostaDeVendaUnica"
        "404":
          description: "Caso a venda não exista retorna a mensagem: Sale not found"
    
    put:
        tags:
        - "Vendas"
        summary: "Atualiza uma venda existente pelo ID"
        description: ""
        operationId: "updateSale"
        consumes:
        - "application/json"
        produces:
        - "application/json"
        parameters:
        - name: "saleId"
          in: "path"
          description: "ID do produto desejado"
          required: true
          type: "integer"
          format: "int64"
        - in: "body"
          name: "body"
          description: "Objeto da venda que será atualizada"
          required: true
          schema:
            $ref: "#/definitions/ModeloDeAtualizacaoDeVenda"
            
        responses:
          "200":
            description: "Retorna o produto atualizado"

    delete:
      tags:
      - "Vendas"
      summary: "Deletar uma venda"
      description: ""
      operationId: "deleteSale"
      produces:
      - "application/json"
      parameters:
      - name: "saleId"
        in: "path"
        description: "ID da venda a ser deletada"
        required: true
        type: "integer"
        format: "int64"
      responses:
        "204":
          description: "Venda excluída com sucesso"
        "404":
          description: "Caso a venda não exista retorna a mensagem: Sale not found"
definitions:
...
  ModeloDeRespostaDeVendaUnica:
    type: "object"
    properties:
      date:
        type: "string"
        example: "2022-04-01T04:48:03.000Z"
      productId:
        type: "integer"
        format: "int64"
      quantity:
        type: "integer"
        format: "int64"
...
  ModeloDeAtualizacaoDeVenda:
    type: "object"
    required:
    - "id"
    - "productId"
    - "quantity"
    properties:
      productId:
        type: "integer"
        format: "int64"
      quantity:
        type: "integer"
        format: "int64"